requirements.txt
Django==1.11.1
django-filter==1.0.3
djangorestframework==3.6.3
djangorestframework-filters==0.10.0
